migrate functions to utils module for easier packaging
